# pre-allocate slices to avoid allocations